separated ProfilePage and Profile component
